[DATA] Merge Data V2 dans Dev

See merge request nguyminh/LO23_Plateforme!35

+++ src/main/java/com/lo23/common/filehandler/FileHandlerStats.java
@@ -20,7 +20,7 @@ public class FileHandlerStats extends FileHandlerInfos
      * @param nbBlocks Nombre de blocks du fichier découpé
      * @param desc Description du fichier
      */
-    public FileHandlerStats (String hash, String title, int size, String type, int nbBlocks, String desc)
+    public FileHandlerStats (String hash, String title, long size, String type, int nbBlocks, String desc)
     {
         super(hash, title, size, type, nbBlocks, desc);
         nbDownloads = 0;
